+++ React/wbooks/src/app/screens/Profile/components/BookSummary/styles.scss
@@ -1,0 +1,19 @@
+@import '../../../../../constants/colors.scss';
+
+.readContainer{
+  height: 316px;
+  padding-top: 35px;
+}
+
+.sectionTitle {
+  color: $special-green;
+  font-weight: bold;
+  font-size: 20px;
+  margin: 0px 0px 24px 0px;
+  line-height: 19px;
+}
+
+.booksContainer {
+  display: flex;
+  padding-left: 37px;
+}
